simplify Hopf model - transfer some code to Network class
